Take 64-bit arithmetic out of the runtime and pass the numeric guard
The guard had 31 findings that arrived between 23 and 30 August while nothing
ran it. The arithmetic ones are rewritten in exact 32-bit form, each checked
against 64-bit arithmetic in a host test:

- poi: the u16-radius distance returns u32; two squares below 2^32 can only
  carry once, and a carry is already outside the radius. A new
  xz_within_radius handles radii past u16 (a vitality circle widened by
  weapon reach) with squares carried as high and low halves.
- vitality: scale_i32_q12 splits the magnitude at the Q12 point instead of
  multiplying and dividing in i64; q12_to_percent does the same for HUD text,
  which lets the playtest drop its u64 decimal formatter and the software
  64-bit divide behind it.
- psx-engine ui: the layout memo's valid mask is two u32 words.

The rest were bitmask sites in the style the code already annotates: the
legacy save checksum decode and two swing-mask parameters get the same
reviewed allow line as their neighbours, and the BSP object pass names its
64-entry chunk instead of borrowing u64::BITS.

+++ engine/examples/editor-playtest/src/playtest_runtime.rs
@@ -1919,7 +1919,7 @@ impl Playtest {
     pub(super) fn find_best_interactable(&mut self) -> Option<usize> {
         let player = self.motor.position();
         let mut best = None;
-        let mut best_distance = u64::MAX;
+        let mut best_distance = u32::MAX;
         for (index, interactable) in INTERACTABLES.iter().enumerate() {
             if !self.point_of_interest_available(interactable)
                 || interactable.room != self.room_index
